+++ server/src/main/java/kg/apps/repository/BalanceRepository.java
@@ -3,5 +3,5 @@ package kg.apps.repository;
 import kg.apps.model.Balance;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BalanceRepository extends JpaRepository<Balance, Long> {
+public interface BalanceRepository extends JpaRepository<Balance, Integer> {
 }
